--- v3/tasks/20250310-v3-vault-factory-v2/index.ts
@@ -1,26 +1,0 @@
-import { VaultFactoryDeployment } from './input';
-import { Task, TaskRunOptions } from '@src';
-import { ethers } from 'hardhat';
-
-/* eslint-disable @typescript-eslint/no-non-null-assertion */
-export default async (task: Task, { force, from }: TaskRunOptions = {}): Promise<void> => {
-  const input = task.input() as VaultFactoryDeployment;
-
-  const vaultFactoryArgs = [
-    input.Authorizer,
-    input.pauseWindowDuration,
-    input.bufferPeriodDuration,
-    input.minTradeAmount,
-    input.minWrapAmount,
-    ethers.utils.keccak256(input.vaultCreationCode),
-    ethers.utils.keccak256(input.vaultExtensionCreationCode),
-    ethers.utils.keccak256(input.vaultAdminCreationCode),
-  ];
-
-  const vaultFactory = await task.deployAndVerify('VaultFactory', vaultFactoryArgs, from, force);
-
-  const vaultAddress = await vaultFactory.getDeploymentAddress(input.salt);
-  if (vaultAddress !== input.targetVaultAddress) {
-    throw Error('Incorrect target address');
-  }
-};
